Code cleanup and restructure
