+++ src/System/Cron/Parser.hs
@@ -12,26 +12,35 @@
 -- Attoparsec parser combinator for cron schedules. See cron documentation for
 -- how those are formatted.
 --
--- > import Data.Attoparsec.Text (parseOnly)
 -- > import System.Cron.Parser
 -- >
 -- > main :: IO ()
--- > main = do
--- >   print $ parseOnly cronSchedule "*/2 * 3 * 4,5,6"
+-- > main = don
+-- >   print $ parseCronSchedule "*/2 * 3 * 4,5,6"
 --
 --------------------------------------------------------------------
-module System.Cron.Parser (cronSchedule,
-                           cronScheduleLoose,
-                           crontab,
-                           crontabEntry) where
+module System.Cron.Parser
+    ( -- * Parsers
+      cronSchedule
+    , cronScheduleLoose
+    , crontab
+    , crontabEntry
+    -- * Convenience Functions
+    , parseCronSchedule
+    , parseCrontab
+    , parseCrontabEntry
+    ) where
 
+-------------------------------------------------------------------------------
 import           Control.Applicative  as Ap
 import           Data.Attoparsec.Text (Parser)
 import qualified Data.Attoparsec.Text as A
 import           Data.Char            (isSpace)
 import           Data.List.NonEmpty   (NonEmpty (..))
 import           Data.Text            (Text)
-import           System.Cron
+-------------------------------------------------------------------------------
+import           System.Cron.Types
+-------------------------------------------------------------------------------
 
 
 -- | Attoparsec Parser for a cron schedule. Complies fully with the standard
@@ -43,6 +52,8 @@ import           System.Cron
 cronSchedule :: Parser CronSchedule
 cronSchedule = cronScheduleLoose <* A.endOfInput
 
+
+-------------------------------------------------------------------------------
 -- | Same as 'cronSchedule' but does not fail on extraneous input.
 cronScheduleLoose :: Parser CronSchedule
 cronScheduleLoose = yearlyP  <|>
@@ -52,6 +63,8 @@ cronScheduleLoose = yearlyP  <|>
                     hourlyP  <|>
                     classicP
 
+
+-------------------------------------------------------------------------------
 -- | Parses a full crontab file, omitting comments and including environment
 -- variable sets (e.g FOO=BAR).
 crontab :: Parser Crontab
@@ -59,6 +72,8 @@ crontab = Crontab <$> A.sepBy lineP (A.char '\n')
   where lineP    = A.skipMany commentP *> crontabEntry
         commentP = A.skipSpace *> A.char '#' *> skipToEOL
 
+
+-------------------------------------------------------------------------------
 -- | Parses an individual crontab line, which is either a scheduled command or
 -- an environmental variable set.
 crontabEntry :: Parser CrontabEntry
@@ -73,15 +88,43 @@ crontabEntry = A.skipSpace *> parser
                           A.skipWhile (\c -> c == ' ' || c == '\t')
                           return $ EnvVariable var val
         commandEntryP = CommandEntry <$> cronScheduleLoose
-                                     <*> (A.skipSpace *> takeToEOL)
+                                     <*> (A.skipSpace *> (CronCommand <$> takeToEOL))
 
----- Internals
+
+-------------------------------------------------------------------------------
+-- Convenience functions
+-------------------------------------------------------------------------------
+
+
+parseCronSchedule :: Text -> Either String CronSchedule
+parseCronSchedule = A.parseOnly cronSchedule
+
+
+-------------------------------------------------------------------------------
+parseCrontab :: Text -> Either String Crontab
+parseCrontab = A.parseOnly crontab
+
+
+-------------------------------------------------------------------------------
+parseCrontabEntry :: Text -> Either String CrontabEntry
+parseCrontabEntry = A.parseOnly crontabEntry
+
+
+-------------------------------------------------------------------------------
+-- Internals
+-------------------------------------------------------------------------------
+
+
 takeToEOL :: Parser Text
 takeToEOL = A.takeTill (== '\n') -- <* A.skip (== '\n')
 
+
+-------------------------------------------------------------------------------
 skipToEOL :: Parser ()
 skipToEOL = A.skipWhile (/= '\n')
 
+
+-------------------------------------------------------------------------------
 classicP :: Parser CronSchedule
 classicP = CronSchedule <$> (minutesP    <* space)
                         <*> (hoursP      <* space)
@@ -90,6 +133,8 @@ classicP = CronSchedule <$> (minutesP    <* space)
                         <*> dayOfWeekP
   where space = A.char ' '
 
+
+-------------------------------------------------------------------------------
 cronFieldP :: Parser CronField
 cronFieldP = stepP  <|>
              listP  <|>
@@ -101,6 +146,7 @@ cronFieldP = stepP  <|>
     stepP         = StepField' <$> stepFieldP
 
 
+-------------------------------------------------------------------------------
 stepFieldP :: Parser StepField
 stepFieldP = do
   f <- baseFieldP
@@ -108,6 +154,7 @@ stepFieldP = do
   mParse (mkStepField f) "invalid stepping" =<< parseInt
 
 
+-------------------------------------------------------------------------------
 neListP :: Parser a -> Parser (NonEmpty a)
 neListP p = coerceNE =<< A.sepBy1 p (A.char ',')
   where
@@ -115,6 +162,8 @@ neListP p = coerceNE =<< A.sepBy1 p (A.char ',')
     coerceNE [_]    = fail "invalid singleton list"
     coerceNE (x:xs) = return $ x :| xs
 
+
+-------------------------------------------------------------------------------
 baseFieldP :: Parser BaseField
 baseFieldP = rangeP <|>
              starP  <|>
@@ -124,10 +173,13 @@ baseFieldP = rangeP <|>
         specificP     = SpecificField' <$> specificFieldP
 
 
+-------------------------------------------------------------------------------
 specificFieldP :: Parser SpecificField
 specificFieldP =
   mParse mkSpecificField "specific field value out of range" =<< parseInt
 
+
+-------------------------------------------------------------------------------
 rangeFieldP :: Parser RangeField
 rangeFieldP = do
   begin <- parseInt
@@ -138,39 +190,61 @@ rangeFieldP = do
   mParse (mkRangeField begin) "start of range must be less than or equal to end" end
 
 
+-------------------------------------------------------------------------------
 yearlyP :: Parser CronSchedule
 yearlyP  = A.string "@yearly"  *> pure yearly
 
+
+-------------------------------------------------------------------------------
 monthlyP :: Parser CronSchedule
 monthlyP = A.string "@monthly" *> pure monthly
 
+
+-------------------------------------------------------------------------------
 weeklyP :: Parser CronSchedule
 weeklyP  = A.string "@weekly"  *> pure weekly
 
+
+-------------------------------------------------------------------------------
 dailyP :: Parser CronSchedule
 dailyP   = A.string "@daily"   *> pure daily
 
+
+-------------------------------------------------------------------------------
 hourlyP :: Parser CronSchedule
 hourlyP  = A.string "@hourly"  *> pure hourly
 
---TODO: must handle a combination of many of these. EITHER just *, OR a list of
+
+-------------------------------------------------------------------------------
 minutesP :: Parser MinuteSpec
 minutesP = mParse mkMinuteSpec "minutes out of range" =<< cronFieldP
 
+
+-------------------------------------------------------------------------------
 hoursP :: Parser HourSpec
 hoursP = mParse mkHourSpec "hours out of range" =<< cronFieldP
 
+
+-------------------------------------------------------------------------------
 dayOfMonthP :: Parser DayOfMonthSpec
 dayOfMonthP = mParse mkDayOfMonthSpec "day of month out of range" =<< cronFieldP
 
+
+-------------------------------------------------------------------------------
 monthP :: Parser MonthSpec
 monthP = mParse mkMonthSpec "month out of range" =<< cronFieldP
 
+
+-------------------------------------------------------------------------------
 dayOfWeekP :: Parser DayOfWeekSpec
 dayOfWeekP = mParse mkDayOfWeekSpec "day of week out of range" =<< cronFieldP
 
+
+-------------------------------------------------------------------------------
 parseInt :: Parser Int
 parseInt = A.decimal
 
+
+-------------------------------------------------------------------------------
 mParse :: (Monad m) => (a -> Maybe b) -> String -> a -> m b
 mParse f msg = maybe (fail msg) return . f
