+++ src/System/Cron/Schedule.hs
@@ -1,8 +1,7 @@
+{-# LANGUAGE CPP                        #-}
 {-# LANGUAGE DeriveFunctor              #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE ScopedTypeVariables        #-}
-{-# LANGUAGE CPP                        #-}
-
 --------------------------------------------------------------------
 -- |
 -- Module      : System.Cron.Schedule
@@ -44,6 +43,8 @@ module System.Cron.Schedule
     , execSchedule
     ) where
 
+
+-------------------------------------------------------------------------------
 #if !MIN_VERSION_base(4,8,0)
 import           Control.Applicative
 #endif
@@ -51,14 +52,19 @@ import           Control.Concurrent
 import           Control.Monad.Except
 import           Control.Monad.Identity
 import           Control.Monad.State
-import           Data.Attoparsec.Text   (parseOnly)
-import           Data.Text              (pack)
+import           Data.Attoparsec.Text       (parseOnly)
+import           Data.Text                  (pack)
 import           Data.Time
-import           System.Cron
-import           System.Cron.Parser
 #if !MIN_VERSION_time(1,5,0)
 import           System.Locale
 #endif
+-------------------------------------------------------------------------------
+import           System.Cron.Internal.Check
+import           System.Cron.Parser
+import           System.Cron.Types
+-------------------------------------------------------------------------------
+
+
 
 readTime' :: TimeLocale -> String -> String -> UTCTime
 #if MIN_VERSION_time(1,5,0)
@@ -67,31 +73,47 @@ readTime' =  parseTimeOrError True
 readTime' = readTime
 #endif
 
-{- Scheduleing monad -}
 
+-------------------------------------------------------------------------------
+-- | Scheduling Monad
 data Job = Job CronSchedule (IO ())
+
+-------------------------------------------------------------------------------
 type Jobs = [Job]
+
 
 instance Show Job where
     show (Job c _) = "(Job " ++ show c ++ ")"
 
+
+-------------------------------------------------------------------------------
 data ScheduleError = ParseError String
                    deriving (Show)
 
+
+-------------------------------------------------------------------------------
 type Schedule = ScheduleT Identity
 
+
+-------------------------------------------------------------------------------
 newtype ScheduleT m a = ScheduleT { unSchedule :: StateT Jobs (ExceptT ScheduleError m) a }
         deriving ( Functor, Applicative, Monad
                  , MonadState Jobs
                  , MonadError ScheduleError
                  )
 
+
+-------------------------------------------------------------------------------
 runSchedule :: Schedule a -> Either ScheduleError (a, [Job])
 runSchedule = runIdentity . runScheduleT
 
+
+-------------------------------------------------------------------------------
 runScheduleT :: ScheduleT m a -> m (Either ScheduleError (a, [Job]))
 runScheduleT = runExceptT . flip runStateT [] . unSchedule
 
+
+-------------------------------------------------------------------------------
 class MonadSchedule m where
     addJob ::  IO () -> String -> m ()
 
@@ -101,13 +123,17 @@ instance (Monad m) => MonadSchedule (ScheduleT m) where
                         Left  e  -> throwError $ ParseError e
                         Right t' -> put $ Job t' a : s
 
-{- Monitoring Engine -}
 
--- | Schedule all of the jobs to run at appropriate intervals. Each 
--- job that is launched gets a scheduling thread to itself. Each 
+-------------------------------------------------------------------------------
+-- Monitoring engine
+-------------------------------------------------------------------------------
+
+
+-- | Schedule all of the jobs to run at appropriate intervals. Each
+-- job that is launched gets a scheduling thread to itself. Each
 -- time a scheduling thread launches a job, the job is forked onto
 -- a new thread. This means that if a job throws an excpetion in IO,
--- its thread will be killed, but it will continue to be scheduled 
+-- its thread will be killed, but it will continue to be scheduled
 -- in the future.
 execSchedule :: Schedule () -> IO [ThreadId]
 execSchedule s = let res = runSchedule s
@@ -115,7 +141,9 @@ execSchedule s = let res = runSchedule s
                         Left  e         -> print e >> return []
                         Right (_, jobs) -> mapM forkJob jobs
 
--- | Start a job-runner thread that runs a job at appropriate 
+
+-------------------------------------------------------------------------------
+-- | Start a job-runner thread that runs a job at appropriate
 -- intervals. Each time it is run, a new thread is forked for it,
 -- meaning that a single exception does not take down the
 -- scheduler.
@@ -125,6 +153,8 @@ forkJob (Job s a) = forkIO $ forever $ do
             threadDelay delay
             when (scheduleMatches s timeAt) (void $ forkIO a)
 
+
+-------------------------------------------------------------------------------
 findNextMinuteDelay :: IO (UTCTime, Int)
 findNextMinuteDelay = do
         now <- getCurrentTime
