+++ test/SpecHelper.hs
@@ -22,7 +22,6 @@ import           Test.Tasty.HUnit      as X
 import           Test.Tasty.QuickCheck as X
 -------------------------------------------------------------------------------
 import           System.Cron           as X
-import           System.Cron.Parser    as X
 -------------------------------------------------------------------------------
