+++ test/System/Test/Cron/Parser.hs
@@ -179,7 +179,7 @@ envSet :: CrontabEntry
 envSet = EnvVariable "FOO" "BAR"
 
 entry :: CrontabEntry
-entry = CommandEntry stars "do stuff"
+entry = CommandEntry stars (CronCommand "do stuff")
 
 stars :: CronSchedule
 stars = everyMinute
