+++ test/System/Test/Cron/Schedule.hs
@@ -6,7 +6,6 @@ module System.Test.Cron.Schedule
 import           Control.Concurrent
 -------------------------------------------------------------------------------
 import           SpecHelper
-import           System.Cron.Schedule
 -------------------------------------------------------------------------------
